Relevant BUGIDs:
Purpose of commit: new feature

Commit summary:
---------------
2008-11-28  Tomas Mraz <t8m@centrum.cz>

        * modules/pam_tally2/pam_tally2.c (tally_check): Fix info format
        to be the same as in pam_tally.

        * configure.in: Add modules/pam_timestamp/Makefile.
        * doc/sag/Linux-PAM_SAG.xml: Include pam_timestamp.xml.
        * doc/sag/pam_timestamp.xml: New.
        * libpam/pam_static_modules.h: Add pam_timestamp static struct.
        * modules/Makefile.am: Add pam_timestamp directory.
        * modules/pam_timestamp/Makefile.am: New.
        * modules/pam_timestamp/README.xml: New.
        * modules/pam_timestamp/hmacsha1.h: New.
        * modules/pam_timestamp/sha1.h: New.
        * modules/pam_timestamp/pam_timestamp.8.xml: New.
        * modules/pam_timestamp/pam_timestamp_check.8.xml: New.
        * modules/pam_timestamp/pam_timestamp.c: New.
        * modules/pam_timestamp/pam_timestamp_check.c: New.
        * modules/pam_timestamp/hmacfile.c: New.
        * modules/pam_timestamp/hmacsha1.c: New.
        * modules/pam_timestamp/sha1.c: New.
        * modules/pam_timestamp/tst-pam_timestamp: New.
        * po/POTFILES.in: Add pam_timestamp sources.
        * po/*.po: Regenerate.
        * po/cs.po: Updated translations.

+++ modules/pam_tally2/pam_tally2.c
@@ -517,8 +517,8 @@ tally_check (tally_t oldcnt, time_t oldtime, pam_handle_t *pamh, uid_t uid,
 #endif
 
         if (!(opts->ctrl & OPT_QUIET)) {
-            pam_info(pamh, _("Account locked due to %hu failed logins"),
-		    tally->fail_cnt);
+            pam_info(pamh, _("Account locked due to %u failed logins"),
+		    (unsigned int)tally->fail_cnt);
         }
 	if (!(opts->ctrl & OPT_NOLOGNOTICE)) {
             pam_syslog(pamh, LOG_NOTICE,
